refactor: remove user affinity sync persistence
