+++ api/src/changeRequests/processChangeRequest.ts
@@ -23,8 +23,6 @@ import processRedirectDto from "./documentProcessing/processRedirectDto";
 import { RedirectDto } from "../dto/RedirectDto";
 import { createSlugChangeRedirect, findSlugReversionRedirect } from "./createSlugChangeRedirect";
 import { isTrackableSlugChange } from "./computePreviousSlugs";
-import processUserAffinityDto from "./documentProcessing/processUserAffinityDto";
-import { UserAffinityDto } from "../dto/UserAffinityDto";
 import processDefaultAffinityDto from "./documentProcessing/processDefaultAffinityDto";
 import { DefaultAffinityDto } from "../dto/DefaultAffinityDto";
 
@@ -98,7 +96,6 @@ export async function processChangeRequest(
             processAuthProviderDto(doc as AuthProviderDto, prevDoc as AuthProviderDto, db),
         [DocType.User]: () => processUserDto(doc as UserDto),
         [DocType.Redirect]: () => processRedirectDto(doc as RedirectDto),
-        [DocType.UserAffinity]: () => processUserAffinityDto(doc as UserAffinityDto, userId),
         [DocType.DefaultAffinity]: () => processDefaultAffinityDto(doc as DefaultAffinityDto),
         [DocType.AutoGroupMappings]: () => {}, // No extra processing required, but needed to be part of the process map for access validation,
     };
