+++ api/src/changeRequests/validateChangeRequest.ts
@@ -16,7 +16,6 @@ import { RedirectDto } from "../dto/RedirectDto";
 import { StorageDto } from "../dto/StorageDto";
 import { AuthProviderDto } from "../dto/AuthProviderDto";
 import { AutoGroupMappingsDto } from "../dto/AutoGroupMappingsDto";
-import { UserAffinityDto } from "../dto/UserAffinityDto";
 import { DefaultAffinityDto } from "../dto/DefaultAffinityDto";
 /**
  * DocType to DTO map
@@ -32,7 +31,6 @@ const DocTypeMap = {
     storage: StorageDto,
     authProvider: AuthProviderDto,
     autoGroupMappings: AutoGroupMappingsDto,
-    userAffinity: UserAffinityDto,
     defaultAffinity: DefaultAffinityDto,
 };
