+++ api/src/changeRequests/validateChangeRequestAccess.ts
@@ -12,7 +12,6 @@ import { _contentBaseDto } from "../dto/_contentBaseDto";
 import { _contentParentDto } from "../dto/_contentParentDto";
 import { GroupDto } from "../dto/GroupDto";
 import { hasTranslateAccessToAllTranslations } from "./hasTranslateAccessToAllTranslations";
-import { userAffinityId } from "../util/userAffinity";
 
 /**
  * Validate a change request against a user's access map
@@ -58,23 +57,6 @@ export async function validateChangeRequestAccess(
         return {
             validated: false,
             error: "Invalid document type - cannot submit Change documents",
-        };
-    }
-
-    // UserAffinity: owner-scoped, not group-scoped. A user may only write their
-    // OWN affinity doc, whose id is derived from their user id. No group ACL
-    // applies. Return early (before the memberOf guard below, which the doc
-    // intentionally can't satisfy — it belongs to no group).
-    if (doc.type === DocType.UserAffinity) {
-        if (!userId || doc._id !== userAffinityId(userId)) {
-            return {
-                validated: false,
-                error: "No access to this affinity document",
-            };
-        }
-        return {
-            validated: true,
-            validatedData: doc,
         };
     }
