+++ api/src/db/schemaUpgrade/README.md
@@ -185,4 +185,4 @@ Backfills the new `CmsView` ACL permission (GitHub #160). CmsView gates CMS-scop
 
 ### v20 — Default affinity ACL + singleton backfill (2026-07-09)
 
-Backfills the CMS-managed "default affinity" recommendation feature (new `DocType.DefaultAffinity`) for existing databases. Adding the doc type itself needs no migration (existing data is unaffected), but two things need establishing on a pre-existing database: (1) `group-super-admins` gets a `DefaultAffinity` ACL entry (View/Edit/Delete/Assign/CmsView) — mirrors the v19 pattern, since without it existing super admins have no permission to create/edit the singleton via the CMS; (2) the singleton doc itself (fixed `_id` = `DEFAULT_AFFINITY_ID`, `api/src/util/userAffinity.ts`) is seeded with an empty profile if absent, so the CMS settings page has something to open immediately rather than only after the first save. Idempotent: the ACL grant is skipped where already present (`insertDoc`, preserving `updatedTimeUtc` like v19); the singleton is only created if missing (`upsertDoc`, which stamps `updatedTimeUtc` itself — the v9 new-doc pattern).
+Backfills the CMS-managed "default affinity" recommendation feature (new `DocType.DefaultAffinity`) for existing databases. Adding the doc type itself needs no migration (existing data is unaffected), but two things need establishing on a pre-existing database: (1) `group-super-admins` gets a `DefaultAffinity` ACL entry (View/Edit/Delete/Assign/CmsView) — mirrors the v19 pattern, since without it existing super admins have no permission to create/edit the singleton via the CMS; (2) the singleton doc itself (fixed `_id` = `DEFAULT_AFFINITY_ID`, `api/src/util/defaultAffinity.ts`) is seeded with an empty profile if absent, so the CMS settings page has something to open immediately rather than only after the first save. Idempotent: the ACL grant is skipped where already present (`insertDoc`, preserving `updatedTimeUtc` like v19); the singleton is only created if missing (`upsertDoc`, which stamps `updatedTimeUtc` itself — the v9 new-doc pattern).
