+++ api/src/socketio.ts
@@ -16,7 +16,6 @@ import { ChangeReqDto } from "./dto/ChangeReqDto";
 import { AccessMap } from "./permissions/permissions.service";
 import configuration, { Configuration } from "./configuration";
 import { JwtUserDetails } from "./auth/authIdentity.service";
-import { UserAffinityDto } from "./dto/UserAffinityDto";
 import { S3Service } from "./s3/s3.service";
 import { WINSTON_MODULE_PROVIDER } from "nest-winston";
 import { Logger } from "winston";
@@ -53,8 +52,8 @@ type ClientConfig = {
     maxUploadFileSize: number;
     maxMediaUploadFileSize?: number;
     accessMap?: AccessMap;
-    /** The caller's own recommendation affinity profile (per-user private). */
-    affinity?: UserAffinityDto;
+    /** CMS-managed cold-start profile for a new client-local affinity store. */
+    defaultAffinity?: Record<string, number>;
 };
 
 /**
@@ -181,11 +180,6 @@ export class Socketio implements OnGatewayInit {
                 );
                 return;
             }
-
-            // UserAffinity docs are per-user private — never fan out over group
-            // rooms. (They also carry no memberOf, so the refGroups guard below
-            // would drop them anyway; this is explicit defense-in-depth.)
-            if (update.type === DocType.UserAffinity) return;
 
             // We are using a socket.io room per document type per group. Change documents are broadcasted to the document-group rooms of the documents they reference.
             // Content documents are broadcasted to their parent document-group rooms.
