+++ api/src/util/defaultAffinity.ts
