+++ app/src/sync.spec.ts
@@ -59,7 +59,6 @@ describe("sync.ts", () => {
             [DocType.Crypto]: [],
             [DocType.AuthProvider]: [],
             [DocType.AutoGroupMappings]: [],
-            [DocType.UserAffinity]: [],
             [DocType.DefaultAffinity]: [],
         });
     });
@@ -131,7 +130,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -158,7 +156,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -185,7 +182,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -217,7 +213,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -245,7 +240,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -309,7 +303,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -345,7 +338,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -381,7 +373,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -429,7 +420,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -465,7 +455,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
