+++ cms/src/sync.spec.ts
@@ -64,7 +64,6 @@ describe("sync.ts", () => {
             [DocType.Crypto]: [],
             [DocType.AuthProvider]: [],
             [DocType.AutoGroupMappings]: [],
-            [DocType.UserAffinity]: [],
             [DocType.DefaultAffinity]: [],
         });
     });
@@ -122,7 +121,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -149,7 +147,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -176,7 +173,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -209,7 +205,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -237,7 +232,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -300,7 +294,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -369,7 +362,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -402,7 +394,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
 
@@ -438,7 +429,6 @@ describe("sync.ts", () => {
                 [DocType.Crypto]: [],
                 [DocType.AuthProvider]: [],
                 [DocType.AutoGroupMappings]: [],
-                [DocType.UserAffinity]: [],
                 [DocType.DefaultAffinity]: [],
             });
