+++ shared/src/types/enum.ts
@@ -11,12 +11,8 @@ export enum DocType {
     Crypto = "crypto",
     AuthProvider = "authProvider",
     AutoGroupMappings = "autoGroupMappings",
-    // Per-user private recommendation affinity profile. Owner-scoped, never
-    // group-synced or /query-able (delivered via authIdentity/clientConfig).
-    UserAffinity = "userAffinity",
-    // CMS-editable global baseline affinity profile (singleton). A normal
-    // group-scoped/permissioned doc (unlike UserAffinity) — cloned into a
-    // first-time user's own UserAffinity scaffold at login (cold start).
+    // CMS-editable global baseline affinity profile (singleton). Delivered at
+    // login to seed a client-local recommendation profile (cold start).
     DefaultAffinity = "defaultAffinity",
 }
